Server/requirements.txt
alembic==1.16.4
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.2.2
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
simple-websocket==1.1.0
SQLAlchemy==2.0.42
typing_extensions==4.14.1
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
